♻️ refactor: integrate tag-team functionality into chat command
- Move TagTeam functionality from standalone command to chat subcommand
- Add 'sidekick chat team' command with same features as tag-team
- Update chat info command to include team agent documentation
- Remove standalone tag-team CLI module and registration
- Update README to reflect new 'chat team' command structure
- Unify all chat-based interactions under single command namespace

BREAKING CHANGE: 'sidekick tag-team chat' is now 'sidekick chat team'
